refactoring: 패키지 변경, 클래스 이름 변경
