+++ app/src/main/java/com/beside153/peopleinside/view/login/nonmember/NonMemberMbtiChoiceActivity.kt
@@ -16,12 +16,12 @@ import com.beside153.peopleinside.util.addBackPressedAnimation
 import com.beside153.peopleinside.util.setCloseActivityAnimation
 import com.beside153.peopleinside.util.setOpenActivityAnimation
 import com.beside153.peopleinside.view.MainActivity
-import com.beside153.peopleinside.view.onboarding.signup.MbtiScreenAdapter
+import com.beside153.peopleinside.view.common.MbtiChoiceScreenAdapter
 import com.beside153.peopleinside.viewmodel.login.nonmember.NonMemberMbtiChoiceViewModel
 
 class NonMemberMbtiChoiceActivity : BaseActivity() {
     private lateinit var binding: ActivityNonMemberMbtiChoiceBinding
-    private val mbtiAdapter = MbtiScreenAdapter(::onMbtiItemClick)
+    private val mbtiAdapter = MbtiChoiceScreenAdapter(::onMbtiItemClick)
     private val mbtiChoiceViewmodel: NonMemberMbtiChoiceViewModel by viewModels()
 
     override fun onCreate(savedInstanceState: Bundle?) {
