+++ app/src/main/java/com/beside153/peopleinside/view/onboarding/signup/SignUpUserInfoFragment.kt
@@ -13,6 +13,7 @@ import com.beside153.peopleinside.R
 import com.beside153.peopleinside.base.BaseFragment
 import com.beside153.peopleinside.databinding.FragmentSignUpUserInfoBinding
 import com.beside153.peopleinside.util.EventObserver
+import com.beside153.peopleinside.view.common.BirthYearBottomSheetFragment
 import com.beside153.peopleinside.viewmodel.onboarding.signup.SignUpUserInfoViewModel
 import com.google.firebase.analytics.FirebaseAnalytics
 import com.google.firebase.analytics.ktx.analytics
@@ -60,7 +61,7 @@ class SignUpUserInfoFragment : BaseFragment() {
         userInfoViewModel.birthYearClickEvent.observe(
             viewLifecycleOwner,
             EventObserver {
-                val bottomSheet = SignUpBottomSheetFragment.newInstance(year)
+                val bottomSheet = BirthYearBottomSheetFragment.newInstance(year)
                 bottomSheet.show(childFragmentManager, bottomSheet.tag)
             }
         )
@@ -107,7 +108,7 @@ class SignUpUserInfoFragment : BaseFragment() {
 
     private fun setFragmentsResultListener() {
         childFragmentManager.setFragmentResultListener(
-            SignUpBottomSheetFragment::class.java.simpleName,
+            BirthYearBottomSheetFragment::class.java.simpleName,
             this
         ) { _, bundle ->
             year = bundle.getInt(YEAR_KEY)
